services/auth/requirements.txt
Django>=4.1,<4.2
djangorestframework
djangorestframework-simplejwt
qrcode
django_otp
pillow